+++ packages/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @workflow/core
 
+## 4.2.0-beta.65
+
+### Patch Changes
+
+- Updated dependencies [[`456c1aa`](https://github.com/vercel/workflow/commit/456c1aa455d9d391a954b25e3d86ee9b06ad2f30), [`11dcb64`](https://github.com/vercel/workflow/commit/11dcb646d33e7a2b251d9388c2c8ecdd6aca73f7)]:
+  - @workflow/world-local@4.1.0-beta.38
+  - @workflow/world@4.1.0-beta.10
+  - @workflow/world-vercel@4.1.0-beta.39
+  - @workflow/errors@4.1.0-beta.18
+
 ## 4.2.0-beta.64
 
 ### Minor Changes