+++ packages/rollup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workflow/rollup
 
+## 4.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.56
+
 ## 4.0.0-beta.21
 
 ### Patch Changes