+++ packages/web-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @workflow/web-shared
 
+## 4.1.0-beta.60
+
+### Patch Changes
+
+- Updated dependencies [[`11dcb64`](https://github.com/vercel/workflow/commit/11dcb646d33e7a2b251d9388c2c8ecdd6aca73f7)]:
+  - @workflow/world@4.1.0-beta.10
+  - @workflow/core@4.2.0-beta.65
+
 ## 4.1.0-beta.59
 
 ### Patch Changes